feat: api endpoints for celebration board with rectified folder/file structuring and naming
