--- src/services/dsmSection/dsmSection.sentimentMeter.services.js
@@ -1,1 +1,0 @@
-cd
